# ENH Speed-up expected mutual information