--- a/doc/whats_new/v1.3.rst
+++ b/doc/whats_new/v1.3.rst
@@ -163,8 +163,8 @@ Changelog
 ......................
 
 - |Efficiency| The computation of the expected mutual information in
-  :func:`adjusted_mutual_info_score` is now faster when the number of labels is large
-  and its memory usage is reduced in general.
+  :func:`metrics.adjusted_mutual_info_score` is now faster when the number
+  of labels is large and its memory usage is reduced in general.
   :pr:`25713` by :user:`Kshitij Mathur <Kshitij68>`,
   :user:`Guillaume Lemaitre <glemaitre>` and
   :user:`Jérémie du Boisberranger <jeremiedbb>`.