Fixes #24254
Closes #24794 
Closes #25575 

This PR proposes an alternative to #24254 and https://github.com/scikit-learn/scikit-learn/pull/24794#issuecomment-1301055896 to speed-up expected mutual information. It avoids creating a large intermediate array and speeds up computations when the number of labels is large while keeping the actual efficiency when the number of labels is small.

```py
import numpy as np
from sklearn.metrics.cluster._expected_mutual_info_fast import expected_mutual_information
from sklearn.metrics.cluster import contingency_matrix

rng = np.random.default_rng(0)
n_samples = 1_000_000
x = rng.integers(low=0, high=5, size=n_samples)  # few labels
y = rng.integers(low=0, high=5, size=n_samples)
#x = np.array([x % 8000 for x in range(10_000)])  # many labels
#y = np.array([x % 7000 for x in range(10_000)])

contingency = contingency_matrix(x, y, sparse=True)
contingency = contingency.astype(np.float64, copy=False)

%timeit expected_mutual_information(contingency, n_samples)
```
On main:
  - few labels: ~350ms
  - many labels: ~20s
This PR:
  - few labels: ~350ms
  - many labels: ~3s

ping @glemaitre @betatim since you reviewed the original PR