@jeremiedbb 2023-02-27
@ogrisel @betatim, in commit [8442447](https://github.com/scikit-learn/scikit-learn/pull/25713/commits/8442447b6b7c41a4af158ac550773c44558b0f53) I used 2 variables, one for the ndarray and one for the memoryview. This way you can more easily see where the memory can be used and where the ndarray must be used. Let me know if you find this clearer.