@glemaitre 2023-03-06
In my original comments, I see that mentioned the following:

It will require changing a line of code in _supervised.py:

```python
contingency = contingency_matrix(
        labels_true, labels_pred, sparse=True, dtype=np.int64
    )
```

which avoid making 3 copy/conversion of data (i.e. int64 -> float64 -> int32) in the current implementation.

Could you have a look at the following?